+++ apps/phoenix_sync_example/config/test.exs
@@ -10,7 +10,7 @@ config :phoenix_sync_example, PhoenixSyncExample.Repo,
   password: "password",
   hostname: "localhost",
   database: "phoenix_sync_example_test#{System.get_env("MIX_TEST_PARTITION")}",
-  port: 55555,
+  port: 54321,
   pool: Ecto.Adapters.SQL.Sandbox,
   pool_size: System.schedulers_online() * 2
 
@@ -33,4 +33,5 @@ config :phoenix_live_view,
 
 config :phoenix_sync,
   env: config_env(),
-  mode: :sandbox
+  mode: :embedded,
+  repo: PhoenixSyncExample.Repo