+++ apps/phoenix_sync_example/lib/phoenix_sync_example/repo.ex
@@ -1,7 +1,5 @@
 defmodule PhoenixSyncExample.Repo do
-  use Phoenix.Sync.Sandbox.Postgres
-
   use Ecto.Repo,
     otp_app: :phoenix_sync_example,
-    adapter: Phoenix.Sync.Sandbox.Postgres.adapter()
+    adapter: Ecto.Adapters.Postgres
 end