+++ apps/plug_sync/config/test.exs
@@ -1,3 +1,11 @@
 import Config
 
-config :phoenix_sync, mode: :sandbox, env: config_env()
+config :phoenix_sync, mode: :embedded, env: config_env(), repo: PlugSync.Repo
+
+config :plug_sync, PlugSync.Repo,
+  username: "postgres",
+  password: "password",
+  hostname: "localhost",
+  database: "phoenix_sync",
+  port: 54321,
+  pool: Ecto.Adapters.SQL.Sandbox