+++ test/phoenix/sync/sandbox/producer_test.exs
@@ -1,5 +1,7 @@
 defmodule Phoenix.Sync.Sandbox.ProducerTest do
   use ExUnit.Case, async: true
 
+  @moduletag :sandbox
+
   doctest Phoenix.Sync.Sandbox.Producer, import: true
 end